backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
pgvector==0.3.3
python-dateutil==2.9.0.post0
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.4.0
redis==5.0.8
tenacity==8.5.0
structlog==24.1.0
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.6.9
jinja2==3.1.4
reportlab==4.2.5
google-generativeai==0.8.1
Pillow==10.4.0
PyMuPDF==1.24.9
email-validator==2.1.0
bcrypt==4.1.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
